Imp: support for article numbers >= 2^31
git-svn-id: http://svn.xananews.techtips.com.br/trunk@345 b371b8be-f639-0410-b043-d2e712c17c34

+++ source/NewsgroupsDialog.pas
@@ -352,7 +352,7 @@ end;
 procedure TdlgNewsgroups.lvNewsgroupsData(Sender: TObject; Item: TListItem);
 var
   st, stx: string;
-  hi, lo: DWORD;
+  hi, lo: Int64;
   isNew: Boolean;
 begin
   if fFiltered then             // Split newgroup line into it's constituents
